+++ client/api.js
@@ -1,0 +1,43 @@
+import store from './store.js';
+
+const base = 'https://conduit.productionready.io/api';
+
+function send({ method, path, data }) {
+	const opts = { method, headers: {} };
+	const { user } = store.get();
+
+	if (data) {
+		opts.headers['Content-Type'] = 'application/json';
+		opts.body = JSON.stringify(data);
+	}
+
+	if (user && user.token) {
+		opts.headers['Authorization'] = `Token ${user.token}`;
+	}
+
+	return fetch(`${base}/${path}`, opts)
+		.then(r => r.text())
+		.then(json => {
+			try {
+				return JSON.parse(json);
+			} catch (err) {
+				return json;
+			}
+		});
+}
+
+export function get(path) {
+	return send({ method: 'GET', path });
+}
+
+export function del(path) {
+	return send({ method: 'DELETE', path });
+}
+
+export function post(path, data) {
+	return send({ method: 'POST', path, data });
+}
+
+export function put(path, data) {
+	return send({ method: 'PUT', path, data });
+}
