+++ client/store.js
@@ -1,0 +1,3 @@
+import { Store } from 'svelte/store.js';
+
+export default new Store();
