+++ routes/article/[slug].html
@@ -46,8 +46,8 @@
 	import Layout from '../_components/Layout.html';
 	import ArticleMeta from './_ArticleMeta.html';
 	import CommentContainer from './_CommentContainer.html';
-	import * as api from '../_api.js';
-	import * as auth from '../_auth.js';
+	import * as api from '../../client/api.js';
+	import * as auth from '../../client/auth.js';
 
 	export default {
 		components: { Layout, ArticleMeta, CommentContainer },
