+++ routes/article/_ArticleMeta.html
@@ -25,7 +25,7 @@
 
 <script>
 	import { goto } from 'sapper/runtime.js';
-	import * as api from '../_api.js';
+	import * as api from '../../client/api.js';
 
 	export default {
 		computed: {
@@ -35,9 +35,8 @@
 		methods: {
 			remove() {
 				const { article } = this.get();
-				const { user } = this.store.get(); // TODO remove
 
-				api.del(`/articles/${article.slug}`, user && user.token).then(response => {
+				api.del(`/articles/${article.slug}`).then(response => {
 					goto('/');
 				});
 			}
