+++ routes/index.html
@@ -31,8 +31,8 @@
 	import Layout from './_components/Layout.html';
 	import MainView from './_components/MainView/index.html';
 	import Tags from './_components/Tags.html';
-	import * as auth from './_auth.js';
-	import * as api from './_api.js';
+	import * as auth from './../client/auth.js';
+	import * as api from './../client/api.js';
 
 	export default {
 		components: { Layout, MainView, Tags },
