+++ routes/login/index.html
@@ -36,7 +36,7 @@
 	import { goto } from 'sapper/runtime.js';
 	import Layout from '../_components/Layout.html';
 	import ListErrors from '../_components/ListErrors.html';
-	import * as auth from '../_auth.js';
+	import * as auth from '../../client/auth.js';
 
 	export default {
 		components: { Layout, ListErrors },
