+++ routes/settings/index.html
@@ -30,7 +30,7 @@
 	import Layout from '../_components/Layout.html';
 	import ListErrors from '../_components/ListErrors.html';
 	import SettingsForm from './_SettingsForm.html';
-	import * as auth from '../_auth.js';
+	import * as auth from '../../client/auth.js';
 
 	export default {
 		components: { Layout, ListErrors, SettingsForm },
@@ -49,10 +49,6 @@
 					this.set({
 						errors: response.errors,
 						inProgress: false
-					});
-
-					this.store.set({
-						user: response.user || this.get('user')
 					});
 				});
 			}
